move some functions from utils/utils.h to engines/utils.h
