+++ src/ppl/nn/engines/common/onnx/if_kernel.cc
@@ -16,7 +16,7 @@
 // under the License.
 
 #include "ppl/nn/engines/common/onnx/if_kernel.h"
-#include "ppl/nn/utils/utils.h"
+#include "ppl/nn/engines/utils.h"
 #include "ppl/nn/utils/generic_cpu_device.h"
 #include "ppl/nn/common/logger.h"
 using namespace std;
