+++ src/ppl/nn/engines/common/onnx/loop_kernel.cc
@@ -18,8 +18,8 @@
 #include "ppl/nn/common/tensor_buffer_info.h"
 #include "ppl/nn/runtime/runtime_impl.h"
 #include "ppl/nn/engines/common/onnx/loop_kernel.h"
+#include "ppl/nn/engines/utils.h"
 #include "ppl/nn/utils/generic_cpu_device.h"
-#include "ppl/nn/utils/utils.h"
 #include "ppl/nn/common/logger.h"
 using namespace std;
 using namespace ppl::common;
