+++ src/ppl/nn/engines/common/ppl/converter_kernel.cc
@@ -16,8 +16,8 @@
 // under the License.
 
 #include "ppl/nn/engines/common/ppl/converter_kernel.h"
+#include "ppl/nn/engines/utils.h"
 #include "ppl/nn/utils/generic_cpu_device.h"
-#include "ppl/nn/utils/utils.h"
 #include "ppl/nn/common/logger.h"
 using namespace std;
 using namespace ppl::common;
